FEATURE Callback: Added back sample payload, hit count refactor

+++ db/migrate/20181127110219_add_hits_to_callback_request.rb
@@ -1,5 +1,5 @@
 class AddHitsToCallbackRequest < ActiveRecord::Migration[5.2]
   def change
-    add_column :callback_requests, :hits, :integer
+    add_column :callback_requests, :hits, :decimal
   end
 end
